fix: ee provider
